Added support for paramaterized class to the ListenerProcess
Added support for a new paramaterized class for ListenerProcess, comprising the original TcpClient, as well as a (string)urlPath, and (string)timeOffset that are necessary on the receiving side of ListenerProcess.

Also made some subtle change to RplCreator to handle situation where a duration is not provided, and therefore the duration is calculated based on TimeOut stamp on the last subtitle.  The edge-case was to handle what happens when the start time of the movie isn't 0, but instead at the 1-hour mark (which sometimes happens in creation of the subtitle file).