Support for multi-RPL loading
Added two new classes (RplPlayoutData and RplLoadInformation)

Added a new static instance of RplLoadInformation that is used by the various commands in the CommandProcess to determine the state of the RPL loads.

Added a few new commands to CommandProcess, including HELP, LIST, SELECT, UNLOAD, and KILL.

Redesigned the functionality for the STOP command (it no longer performs a TerminateLease, that is expected to be done by the KILL command now).

Updated the code for the "TIME" command to use the  RplLoadInformation data.

Added code for LIST and SELECT commands.

Put in stub-holders for UNLOAD, KILL, and HELP for now.

Redesigned the functionality of DoCommandTime to support the multi-RPL and RplLoadInformation design.

Redesigned the functionality of DoCommandLoad to support the multi-RPL and RplLoadInformation design.