Minor update
Small improvements to the way DoCommandSelect is handled, including adding a pause to the current playout if a new select is made (even if that select fails).

Added try/catch logic to the TIME command attempt, as there is a bug that occurs if you do a backspace while entering the TIME command.