Fixes #17 (KILL command needs working functionality)
Added a static bool called KillCommandReceived, which is initialized to false, and reset to false on every successful ListenerClient connection.   The KILL command performs a DoCommandStop() and then a DoCommandKill() now, which simply sets the boolean flag to true (triggering the control logic in the listener process).

Seems to work the way we want it to.