Fixes #23 (RplCreator needs to create UNIQUE PlayoutId)
Added some new code to support a different way to create the PlayoutId uint32 number, which is created at time of creation of the ResourcePresenationList object, and the PlayoutId is supposed to be UNIQUE number between all of the loaded RPLs.

The old way of creating the number created a random value between 10000000 and 99999999.  This new usage creates an int32 from a GetHashCode() call based on the filename being passed in, and then boxes the int32 in to a uint32.  At the very least, this has the added advantage of creating the SAME PlayoutId every time the same filename is loaded, which is kind of what we want as well.