More improvements to the RMS_Proofing project
Added a new form of the RootMeanSquare method to allow for a number of samples to calculate against that isn't the entire length of the supplied array.  Only implemented for the method that takes a float array so far.

On the main form, added some defines to more easily control behavior so I don't have to comment stuff out.  I added three different defines (DSP_USES_SIGNALS, DSP_PLAYS_AUDIO, and DSP_FORCES_FORM_UPDATE).

Also on the main form, I added some new logic to detect when we are on the last sample window and take advantage of the changes I made to RootMeanSquare() method above.

Also on the main form, I added some logic to show the FrameWindowIndex in the RMS ListBox, and added a new checkbox that supports reversing that data (so that you see the latest frame on the top of the list).

Also experimented with the OffsetSampleProvider class just to see how it works, but commented out all of the optional items so that it still behaves as a normal SampleProvider.