Fix for NetworkStream.Flush
Added an extension method for NetworkStream to clear anything out of the stream, so that we can make sure when we are sending Request and receiving Response that we aren't getting any garbage.

Added a pragma rule for three warning messages in AcspTimelineExtension.cs (which we haven't implemented yet)

Cleaned up another warning error.

Added more documentation to various methods.

Changed the Log level on some of the messages to Log.Verbose