Added two catch blocks to address issue 44
While further researching the issue, discovered that in certain conditions we can run in to a race condition where we make it past the validation code and get directly in to ListenerProcess' code where it starts trying to send messages to the ACS (Announce, SetLease, and GetStatus).   Around this point, as it is starting that processing, the OTHER thread is finally successfully cleaning up the ACS connection, which includes disposing and nulling the NetworkStream object "stream", the TcpClient object "ListenerClient", and the AcspLeaseTimer object "leaseTimer".      Once those three objects are disposed and nulled (but remember, they're static objects, so they stick around), then any further attempt to access them at that point results in a NullReferenceException or an ObjectDisposedException (depending where in the race condition we are when it occurs).

Resolves #44