General refactoring / renaming of some local variables
Mostly just a refactoring commit.

Also moved the stop/dispose/null of the AcspLeaseTimer object from the bottom of the cleanup method to the top (just so that if some kind of exception
occurs, that it is being handled first).

Added two new Exception classes for handling custom ACSP-related exceptions.