Fixes for issue #44
Resolves #44

Added new logic in ListenerProcess to perform a ProcessAnnounceRequest() call prior to any killing of existing connections.  In other words, we want to PROVE it is an ACS device that is connected to us before we start cleaning up a previous connection.

In order to facilitate that logic, I had to add some code that creates a temporary TcpClient and NetworkStream object, and changed the signature of ProcessAnnounceRequest() to take a NetworkStream object as a method parameter.  I use the temporary NetworkStream object for all communication in the ProcessAnnounceRequest() method, and if the method succeeds (a valid ACS response is provided and the RequestId matches what was sent) then those temporary TcpClient and NetworkStream objects are assigned over to the static members in the Program class.   If the method fails, I throw a custom AcspAnnounceException, which is handled in ListenerProcess, slightly differently than the rest of the exceptions we anticipate.

As part of that logic change, I felt I needed to add the rplReloadInfo logic to DoAcsConnectionCleanup (since we do not perform that method in the finally clause anymore).

Also did some local variable refactoring/renaming as needed.

Also added some comments/documentation for methods that didn't have it