Changes to support validation of resource file in RPL
Fixes #59

Added new static "SavedLocalAddress" that is captured during the OnAccept call (but only when accepting a connection on the ACS port)

Added new static method in Program class called ValidateSubtitleFromUrlString(), which performs all of the necessary validation steps to ensure that the resource file specified in the RPL is both present, and in expected format.

Minor changes to RplCreator, moving the logic that deserializes the CineCanvas file to the SubtitleReel class.  Jon showed me a different approach that I'm trying out, where we use a static method in the class instead of using the constructor.