Fixes for issue #46
Fixes #46

I've added code throughout all of the Process* methods to handle a situation where ListenerStream is null. I perform a null check right before the ListenerStream.Write call (sending the outgoing message request), and if it fails the null check, then I throw a ArgumentNullException (true, its not really an "argument", its a static variable - but this is the exception that applies). As a result of this exception, I put corresponding catch statement for it in both ListenerProcess and CommandProcess.

Also, as part of the work effort, I upgraded my projects to .Net framework 4.7.2, and upgraded all of my 3rd party NuGet packages to support that as well.