Further fixes for issue #63
Related to issue #63, renamed AcsConnectedDelegate and AcsConnectedEventArgs to AcsConnectionStatusDelegate and AcsConnectionStatusEventArgs to provide a more generic delegate and set of events related to AcsConnectionStatus (connected or disconnected).

Added a new AcsConnectionInfo class type that will hold all of the connection-related information (IPAddress, connection status, etc).

Added two simple subscriptions in program.cs to the OnConnect and OnDisconnect events , just basically as proof of concept for the functionality.