Some changes to support issue #67
Related to #67

Made changes to the DoCommandLoad() method signature, removing the editrate as an input parameter, and giving the DoCommandTime method the responsibility for getting that data once inside the method.  This is how it should have been from the start, really.   Now, with the changes, a consumer of AcsListener class can call this method appropriately, with only the desired timecode passed in to the method. DoCommandTime() does all the rest of the work.

Modified the two calls to DoCommandTime, removing the code related to getting an RplPlayoutData object, and data-checking the EditRate parameter.  Now that DoCommandTime() does all of the work for this, it didn't make sense to redundantly do this one step up in the call stack.